Renaming and refactoring based on peer-review:
https://perfectsense.atlassian.net/browse/BSP-1566?focusedCommentId=1867
20&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpan
el#comment-186720
