+++ src/main/webapp/base/All.less
@@ -15,8 +15,8 @@
 @import "main/CreativeWorkMain.less";
 @import "main/Main.less";
 @import "main/GalleryMain.less";
-@import "main/SearchMain.less";
-@import "main/SearchMainResult.less";
+@import "search/SearchMain.less";
+@import "search/SearchResult.less";
 
 @import "GalleryIntro.less";
 @import "GallerySlide.less";
