+++ src/main/webapp/base/form/Select.hbs
@@ -7,7 +7,7 @@
     {{#defineElement "select" noWith=true}}
         <select
             class="{{elementName}}"
-            name="{{id}}"
+            name="{{name}}"
             id="{{id}}"
             {{#if required}} required{{/if}}
             {{#if autosubmit}} data-bsp-autosubmit{{/if}}
