--- src/main/webapp/base/main/SearchMainResult.less
@@ -1,16 +1,0 @@
-.SearchMainResult {
-    &:extend(.SearchMainResult all);
-
-    &-items {
-        &:extend(.SearchMainResult-items all);
-    }
-
-    &-loadMore {
-        &:extend(.SearchMainResult-loadMore all);
-    }
-
-    &-loadMore-button {
-        &:extend(.SearchMainResult-loadMore-button all);
-        &[disabled] { }
-    }
-}
