+++ src/main/webapp/base/search/SearchMain.less
@@ -1,18 +1,16 @@
-@import 'CreativeWorkMain.less';
-
 .SearchMain {
-    &:extend(.CreativeWorkMain all);
+    &:extend(.SearchMain all);
 
     &-header {
         &:extend(.SearchMain-header all);
     }
 
     &-headline {
-        &:extend(.CreativeWorkMain-headline all);
+        &:extend(.SearchMain-headline all);
     }
 
-    &-byline {
-        &:extend(.CreativeWorkMain-byline all);
+    &-subHeadline {
+        &:extend(.SearchMain-subHeadline all);
     }
 
     &-filter {
