+++ src/main/webapp/base/search/SearchResult.hbs
@@ -1,4 +1,4 @@
-{{#defineBlock "SearchMainResult"}}
+{{#defineBlock "SearchResult"}}
 
     {{#defineElement "items"}}
         <div class="{{elementName}}">{{this}}</div>
