+++ src/main/webapp/main.js
@@ -19,11 +19,11 @@ import CommentEntry from "community/commenting/CommentEntry.js";
 import Commenting from "community/commenting/Commenting.js";
 
 import Gallery from "base/main/GalleryMain.js";
-import { SearchMain } from "base/main/SearchMain.js";
+import { SearchMain } from "base/search/SearchMain.js";
 
 $( document ).ready(function() {
 
-    // Search Results Main
+    // Search Main
     let $searchMain = $('.SearchMain')
     if ($searchMain.length) {
         new SearchMain($searchMain, { })
