+++ styleguide/block/search/_search-main-last-result-response.json
@@ -1,6 +1,6 @@
 {
     "_wrapper": false,
-    "_template": "base/main/SearchMainResult",
+    "_template": "base/search/SearchResult",
 
     "items": {
         "_template": "base/util/Concatenated",
@@ -13,9 +13,5 @@
                 "html": "<div>{{words(3)}}</div>"
             }
         ]
-    },
-
-    "loadMore": true,
-    "loadMoreAction": "_search-main-last-result-response?q=keywords&offset={{number([1,10])}}&limit=3",
-    "loadMoreLabel": "Load More (page 2)"
+    }
 }
