+++ styleguide/form/select/disabled-select.json
@@ -2,6 +2,7 @@
     "_template": "base/form/Select",
 
     "id": "selectId-123",
+    "name": "selectName-123",
     "label": "Choose an option",
     "required": true,
     "disabled": true,
